Split constraint module
To prepare for the implementation of more constraints, the constraint
module was split into 2 categories: "composite" (containing the
CompositeConstraint and DynamicConstraint) and "irreducible" (containing
all other constraint). The idea is that, in the future, reducible
constraints (such as thermo, killer, etc.) are put into a "reducible"
module.
